Phase 1: bake models, clear clippy, format, pin toolchain
- Bake the HOR-score + k-predictor random forests into the binary
  via include_bytes!. Binary goes from ~5 MB to ~26 MB but is now
  fully self-contained — no data files to install or discover.
  CLI flags --hor-model / --k-model still override with a path.
- Drop the now-unused [models] table from classifier.toml and the
  ModelsCfg struct from ClassifierConfig.
- Clear all clippy warnings under -D warnings (23 errors → 0):
  abs_diff, byte-string literals, iter::enumerate, unwrap-after-
  is_some, useless casts, doc-list indentation, unused lifetimes.
- One-time cargo fmt over the tree so cargo fmt --check is clean.
- Pin Rust 1.95 in rust-toolchain.toml (rustfmt + clippy components).
- Polish Cargo.toml: repository/homepage/documentation/keywords/
  categories. Add exclude to keep test_data, ground_truth, tools,
  docs, tests out of any future `cargo package` artefact.

Verification: 47 unit tests pass, integration_smoke passes,
cargo clippy clean, cargo fmt --check clean. Smoke verdicts on
test_data/smoke/ unchanged (tandem_pure → tandem, hor_k3 → hor
k=3 founder=100 tile=300, hor_k5 → hor k=5 founder=150 tile=750).

+++ src/monomer_model.rs
@@ -98,11 +98,7 @@ pub fn score_candidate(
             continue;
         }
         let sim = mean_adjacent_similarity(&blocks, cfg.max_pairs_sampled);
-        if best
-            .as_ref()
-            .map(|(_, b, _)| sim > *b)
-            .unwrap_or(true)
-        {
+        if best.as_ref().map(|(_, b, _)| sim > *b).unwrap_or(true) {
             best = Some((offset, sim, blocks));
         }
     }
@@ -248,7 +244,7 @@ impl<'a> Iterator for CanonicalKmers<'a> {
                     }
                 };
                 fwd = (fwd << 2) | f;
-                rc = (rc >> 2) | ((3 ^ f as u64) << (2 * (self.k - 1)));
+                rc = (rc >> 2) | ((3 ^ f) << (2 * (self.k - 1)));
             }
             if !ok {
                 continue;
